TW-396: Set size for appbar
