+++ lib/pages/forward/forward_view.dart
@@ -17,6 +17,7 @@ import 'package:linagora_design_flutter/colors/linagora_ref_colors.dart';
 
 class ForwardView extends StatefulWidget {
   final ForwardController controller;
+
   const ForwardView(this.controller, {super.key});
 
   @override
@@ -37,7 +38,8 @@ class _ForwardViewState extends State<ForwardView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: PreferredSize(
-        preferredSize: Size.fromHeight(ForwardViewStyle.preferredAppBarSize),
+        preferredSize:
+            Size.fromHeight(ForwardViewStyle.preferredAppBarSize(context)),
         child: _buildAppBarForward(context),
       ),
       body: SingleChildScrollView(
@@ -108,63 +110,71 @@ class _ForwardViewState extends State<ForwardView> {
 
   Widget _buildAppBarForward(BuildContext context) {
     return AppBar(
-      toolbarHeight: 64,
+      toolbarHeight: ForwardViewStyle.preferredAppBarSize(context),
       surfaceTintColor: Colors.transparent,
       leadingWidth: double.infinity,
-      leading: Row(
-        children: [
-          TwakeIconButton(
-            tooltip: L10n.of(context)!.back,
-            icon: Icons.arrow_back,
-            onPressed: () {
-              Matrix.of(context).shareContent = null;
-              context.pop();
-            },
-            paddingAll: 8.0,
-            margin: const EdgeInsets.symmetric(vertical: 12.0),
-          ),
-          const SizedBox(width: 8.0),
-          isSearchBarShow
-              ? Padding(
-                  padding: const EdgeInsets.only(top: 8.0),
-                  child: TextField(
-                    autofocus: true,
-                    maxLines: 1,
-                    buildCounter: (
-                      BuildContext context, {
-                      required int currentLength,
-                      required int? maxLength,
-                      required bool isFocused,
-                    }) =>
-                        const SizedBox.shrink(),
-                    maxLength: 200,
-                    cursorHeight: 26,
-                    scrollPadding: const EdgeInsets.all(0),
-                    decoration: InputDecoration(
-                      isCollapsed: true,
-                      hintText: "...",
-                      hintStyle:
-                          Theme.of(context).textTheme.bodyLarge?.copyWith(
-                                color: LinagoraRefColors.material().neutral[60],
-                              ),
-                    ),
-                  ),
-                )
-              : Text(
-                  L10n.of(context)!.forwardTo,
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                        color: Theme.of(context).colorScheme.onSurface,
-                        letterSpacing:
-                            ChatAppBarTitleStyle.letterSpacingRoomName,
+      leading: Padding(
+        padding: ForwardViewStyle.paddingItemAppbar(context),
+        child: Row(
+          children: [
+            TwakeIconButton(
+              tooltip: L10n.of(context)!.back,
+              icon: Icons.arrow_back,
+              onPressed: () {
+                Matrix.of(context).shareContent = null;
+                context.pop();
+              },
+              paddingAll: 8.0,
+              margin: const EdgeInsets.symmetric(vertical: 12.0),
+            ),
+            const SizedBox(width: 8.0),
+            isSearchBarShow
+                ? Padding(
+                    padding: const EdgeInsets.only(top: 8.0),
+                    child: TextField(
+                      autofocus: true,
+                      maxLines: 1,
+                      buildCounter: (
+                        BuildContext context, {
+                        required int currentLength,
+                        required int? maxLength,
+                        required bool isFocused,
+                      }) =>
+                          const SizedBox.shrink(),
+                      maxLength: 200,
+                      cursorHeight: 26,
+                      scrollPadding: const EdgeInsets.all(0),
+                      decoration: InputDecoration(
+                        isCollapsed: true,
+                        hintText: "...",
+                        hintStyle: Theme.of(context)
+                            .textTheme
+                            .bodyLarge
+                            ?.copyWith(
+                              color: LinagoraRefColors.material().neutral[60],
+                            ),
                       ),
-                ),
-        ],
+                    ),
+                  )
+                : Text(
+                    L10n.of(context)!.forwardTo,
+                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                          color: Theme.of(context).colorScheme.onSurface,
+                          letterSpacing:
+                              ChatAppBarTitleStyle.letterSpacingRoomName,
+                        ),
+                  ),
+          ],
+        ),
       ),
       actions: [
-        TwakeIconButton(
-          icon: Icons.search,
-          onPressed: () {},
-          tooltip: L10n.of(context)!.search,
+        Padding(
+          padding: ForwardViewStyle.paddingItemAppbar(context),
+          child: TwakeIconButton(
+            icon: Icons.search,
+            onPressed: () {},
+            tooltip: L10n.of(context)!.search,
+          ),
         ),
       ],
       bottom: PreferredSize(
