+++ lib/pages/new_private_chat/new_private_chat_view.dart
@@ -1,6 +1,7 @@
 import 'package:fluffychat/pages/new_private_chat/new_private_chat.dart';
 import 'package:fluffychat/pages/new_private_chat/widget/expansion_list.dart';
-import 'package:fluffychat/pages/new_private_chat/widget/search_contact_appbar.dart';
+import 'package:fluffychat/pages/new_private_chat/widget/new_private_appbar.dart';
+import 'package:fluffychat/pages/new_private_chat/widget/new_private_appbar_style.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 
@@ -13,8 +14,10 @@ class NewPrivateChatView extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: PreferredSize(
-        preferredSize: const Size.fromHeight(64),
-        child: SearchContactAppBar(
+        preferredSize: Size.fromHeight(
+          NewPrivateAppBarStyle.appbarHeight(context),
+        ),
+        child: NewPrivateAppBar(
           title: L10n.of(context)!.newChat,
           searchContactsController: controller.searchContactsController,
         ),
