+++ lib/pages/new_private_chat/widget/new_private_appbar_style.dart
@@ -1,15 +1,9 @@
 import 'package:fluffychat/utils/responsive/responsive_utils.dart';
 import 'package:flutter/cupertino.dart';
 
-class ForwardViewStyle {
-  static double preferredAppBarSize(BuildContext context) =>
+class NewPrivateAppBarStyle {
+  static double appbarHeight(BuildContext context) =>
       ResponsiveUtils().isMobile(context) ? 104 : 64;
-
-  static double get paddingBody => 16.0;
-
-  static double get bottomBarHeight => 120.0;
-
-  static double get iconSendSize => 40.0;
 
   static EdgeInsetsDirectional paddingItemAppbar(BuildContext context) =>
       EdgeInsetsDirectional.only(
